Introduce `fetchesMessages` to `AccountAttributes`
Create the field `fetchesMessages` in the class `AccountAttributes`,
add the appropriate parameters to the constructor, and implement a
'getting'.

The functionality for a device to retrieve its own messages is already
implemented on the server but we have no way of accessing it.

Alongside this minimal change is a collection of modifications so that
methods that make use of `AccountAttributes` still do so correctly.

Signed-off-by: Alex Melbourne <alex.melbourne@protonmail.com>